+++ src/app.js
@@ -6,17 +6,17 @@ import { CurrentUserProvider } from 'State/current-user';
 import { APIContextProvider } from 'State/api';
 import { UserPrefsProvider } from 'State/user-prefs';
 import { DevTogglesProvider } from 'State/dev-toggles';
+import { NotificationsProvider } from 'State/notifications';
 import OfflineNotice from 'State/offline-notice';
 import SuperUserBanner from 'Components/banners/super-user';
 
 import ErrorBoundary from './presenters/includes/error-boundary';
-import { Notifications } from './presenters/notifications';
 import Router from './presenters/pages/router';
 
 const App = () => (
   <ErrorBoundary fallback="Something went very wrong, try refreshing?">
     <BrowserRouter>
-      <Notifications>
+      <NotificationsProvider>
         <UserPrefsProvider>
           <DevTogglesProvider>
             <AnalyticsContext context={{ groupId: '0' }}>
@@ -32,7 +32,7 @@ const App = () => (
             </AnalyticsContext>
           </DevTogglesProvider>
         </UserPrefsProvider>
-      </Notifications>
+      </NotificationsProvider>
     </BrowserRouter>
   </ErrorBoundary>
 );
